Refactor Connection Adapter. Extract logic to read settings and to build occupancy into it's own class
