+++ app/src/main/java/io/github/imsmobile/fahrplan/adapter/ConnectionAdapter.java
@@ -1,7 +1,6 @@
 package io.github.imsmobile.fahrplan.adapter;
 
 import android.content.Context;
-import android.content.SharedPreferences;
 import android.support.annotation.NonNull;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -24,7 +23,8 @@ import java.util.Locale;
 import ch.schoeb.opendatatransport.model.Connection;
 import ch.schoeb.opendatatransport.model.Journey;
 import ch.schoeb.opendatatransport.model.Section;
-import io.github.imsmobile.fahrplan.JourneyPrinter;
+import io.github.imsmobile.fahrplan.output.JourneyPrinter;
+import io.github.imsmobile.fahrplan.output.Occupancy;
 import io.github.imsmobile.fahrplan.R;
 import io.github.imsmobile.fahrplan.SearchResultActivity;
 
@@ -34,9 +34,12 @@ public class ConnectionAdapter extends BaseAdapter {
     private final Context context;
     private final List<Connection> connections;
     private static final DateFormat DF = DateFormat.getTimeInstance(DateFormat.SHORT);
+    private final Occupancy occupancy;
+
     public ConnectionAdapter(Context context, List<Connection> connections) {
         this.context = context;
         this.connections = connections;
+        this.occupancy = new Occupancy(context);
     }
 
     @Override
@@ -82,27 +85,11 @@ public class ConnectionAdapter extends BaseAdapter {
         setViewText(view, R.id.item_to, connection.getTo().getStation().getName());
         setViewText(view, R.id.departure_time, formatTimestamp(connection.getFrom().getDepartureTimestamp()));
         setViewText(view, R.id.duration, prettifyDuration(connection.getDuration()));
-        setViewText(view, R.id.occupancy, buildOccupancy(connection));
+        setViewText(view, R.id.occupancy, occupancy.buildOccupancy(connection));
         setViewText(view, R.id.arrival_time, formatTimestamp(Long.parseLong(connection.getTo().getArrivalTimestamp())));
         setViewText(view, R.id.journey_stops, buildVehicles(connection));
     }
 
-    private String buildOccupancy(Connection connection) {
-        StringBuilder occupancy = new StringBuilder();
-        if((connection.getCapacity1st() != null) && Boolean.valueOf(getSettings(context.getString(R.string.setting_classes_first), String.valueOf(true)))) {
-            occupancy.append(" 1.");
-            occupancy.append((char) (getLowOccupanyIcon() + connection.getCapacity1st().intValue()*2));
-        }
-        if((connection.getCapacity2nd() != null) && Boolean.valueOf(getSettings(context.getString(R.string.setting_classes_second), String.valueOf(true)))) {
-            occupancy.append(" 2.");
-            occupancy.append((char) (getLowOccupanyIcon() + connection.getCapacity2nd().intValue()*2));
-        }
-        return occupancy.toString();
-    }
-
-    private char getLowOccupanyIcon() {
-        return context.getString(R.string.icon_occupancy).charAt(0);
-    }
 
     @NonNull
     private String buildVehicles(Connection connection) {
@@ -153,8 +140,4 @@ public class ConnectionAdapter extends BaseAdapter {
         fromTextView.setText(str);
     }
 
-    private String getSettings(String key, String defaultValue) {
-        SharedPreferences preferences = context.getSharedPreferences(context.getString(R.string.setting_name), Context.MODE_PRIVATE);
-        return preferences.getString(key, defaultValue);
-    }
 }
